Added tests, cleaner architecture, the works

+++ LxfmlSharp.Core/Components/Bone.cs
@@ -8,11 +8,11 @@ public sealed class Bone
     public string? Uuid { get; init; }
 
     // 3x4 matrix row-major: [r00,r01,r02,tx, r10,r11,r12,ty, r20,r21,r22,tz]
-    public required double[] Transformation3x4 { get; init; }
+    public required float[] Transformation3x4 { get; init; }
 
     public bool IsValid =>
         Transformation3x4 is { Length: 12 }
-        && Array.TrueForAll(Transformation3x4, d => !double.IsNaN(d) && !double.IsInfinity(d));
+        && Array.TrueForAll(Transformation3x4, d => !float.IsNaN(d) && !float.IsInfinity(d));
 
     [DebuggerBrowsable(DebuggerBrowsableState.Never)]
     private string DebuggerDisplay =>
